atoms

+++ modular_skyrat/modules/modular_items/lewd_items/code/lewd_structures/bdsm_furniture.dm
@@ -14,14 +14,19 @@
 	var/unwrapped = 0
 	w_class = WEIGHT_CLASS_HUGE
 
-/obj/item/bdsm_bed_kit/wrench_act(mob/living/user, obj/item/tool)
-	. = ..()
-	to_chat(user, span_notice("You fasten the frame to the floor and begin to inflate the latex pillows..."))
-	if(tool.use_tool(src, user, 8 SECONDS, volume=50))
-		to_chat(user, span_notice("You assemble the bdsm bed."))
-		var/obj/structure/bed/bdsm_bed/C = new
-		C.loc = loc
-		qdel(src)
+/obj/item/bdsm_bed_kit/attackby(obj/item/P, mob/user, params) //constructing a bed here.
+	add_fingerprint(user)
+	if(istype(P, /obj/item/wrench))
+		if (!(item_flags & IN_INVENTORY))
+			to_chat(user, span_notice("You fasten the frame to the floor and begin to inflate the latex pillows..."))
+			if(P.use_tool(src, user, 8 SECONDS, volume=50))
+				to_chat(user, span_notice("You assemble the bdsm bed."))
+				var/obj/structure/bed/bdsm_bed/C = new
+				C.loc = loc
+				qdel(src)
+			return
+	else
+		return ..()
 
 /obj/structure/bed/bdsm_bed/post_buckle_mob(mob/living/M)
 	density = TRUE
@@ -33,15 +38,19 @@
 	//Set them back down to the normal lying position
 	M.pixel_y = M.base_pixel_y + M.body_position_pixel_y_offset
 
-/obj/structure/bed/bdsm_bed/wrench_act(mob/living/user, obj/item/tool)
-	. = ..()
-	to_chat(user, span_notice("You begin unfastening the frame of bdsm bed and deflating the latex pillows..."))
-	if(tool.use_tool(src, user, 8 SECONDS, volume=50))
-		to_chat(user, span_notice("You disassemble the BDSM bed."))
-		var/obj/item/bdsm_bed_kit/C = new
-		C.loc = loc
-		unbuckle_all_mobs()
-		qdel(src)
+/obj/structure/bed/bdsm_bed/attackby(obj/item/P, mob/user, params) //deconstructing a bed. Aww(
+	add_fingerprint(user)
+	if(istype(P, /obj/item/wrench))
+		to_chat(user, span_notice("You begin unfastening the frame of bdsm bed and deflating the latex pillows..."))
+		if(P.use_tool(src, user, 8 SECONDS, volume=50))
+			to_chat(user, span_notice("You disassemble the BDSM bed."))
+			var/obj/item/bdsm_bed_kit/C = new
+			C.loc = loc
+			unbuckle_all_mobs()
+			qdel(src)
+		return
+	else
+		return ..()
 
 /obj/structure/bed/bdsm_bed/Destroy()
 	. = ..()
@@ -121,7 +130,7 @@
 	// Let's make sure that the X-Stand is in the correct state
 	if(stand_state == "open")
 		toggle_mode(user)
-	var/mob/living/M = buckled_mob
+	var/mob/living/M = unbuckle_mob(buckled_mob)
 	if(M)
 		if(M != user)
 			if(!do_after(user, 5 SECONDS, M)) // Timer for unbuckling one mob with another mob
@@ -141,7 +150,6 @@
 			var/mob/living/L = M.pulledby
 			L.set_pull_offsets(M, L.grab_state)
 		toggle_mode(user)
-	unbuckle_mob(buckled_mob)
 	return M
 
 // Handler for attempting to buckle a mob into a X-Stand
@@ -264,23 +272,30 @@
 	var/unwrapped = 0
 	w_class = WEIGHT_CLASS_HUGE
 
-/obj/item/x_stand_kit/wrench_act(mob/living/user, obj/item/tool)
-	. = ..()
-	to_chat(user, span_notice("You begin fastening the frame to the floor."))
-	if(tool.use_tool(src, user, 8 SECONDS, volume=50))
-		to_chat(user, span_notice("You assemble the x-stand."))
-		var/obj/structure/chair/x_stand/C = new
-		C.loc = loc
-		qdel(src)
-	return TRUE
+/obj/item/x_stand_kit/attackby(obj/item/P, mob/user, params) //constructing a bed here.
+	add_fingerprint(user)
+	if(istype(P, /obj/item/wrench))
+		if (!(item_flags & IN_INVENTORY))
+			to_chat(user, span_notice("You begin fastening the frame to the floor."))
+			if(P.use_tool(src, user, 8 SECONDS, volume=50))
+				to_chat(user, span_notice("You assemble the x-stand."))
+				var/obj/structure/chair/x_stand/C = new
+				C.loc = loc
+				qdel(src)
+			return
+	else
+		return ..()
 
-/obj/structure/chair/x_stand/wrench_act(mob/living/user, obj/item/I)
-	. = ..()
-	to_chat(user, span_notice("You begin unfastening the frame of x-stand..."))
-	if(I.use_tool(src, user, 8 SECONDS, volume=50))
-		to_chat(user, span_notice("You disassemble the x-stand."))
-		var/obj/item/x_stand_kit/C = new
-		C.loc = loc
-		unbuckle_all_mobs()
-		qdel(src)
-	return TRUE
+/obj/structure/chair/x_stand/attackby(obj/item/P, mob/user, params) //deconstructing a bed. Aww(
+	add_fingerprint(user)
+	if(istype(P, /obj/item/wrench))
+		to_chat(user, span_notice("You begin unfastening the frame of x-stand..."))
+		if(P.use_tool(src, user, 8 SECONDS, volume=50))
+			to_chat(user, span_notice("You disassemble the x-stand."))
+			var/obj/item/x_stand_kit/C = new
+			C.loc = loc
+			unbuckle_all_mobs()
+			qdel(src)
+		return
+	else
+		return ..()
